Create workingThreadMain.cc
Implemented pseudoProcess class that derives from Thread. 
Constructor takes in the name, priority, and vector of strings. 
Action creates an argv from the vector of strings and then forks and execs argv (the fork is so that the main thread does not die).

This seems to be working so far. Just need to implement it with the shell code now.  Maybe move pseudoProcess into the thread.h file. 

Note: If you want them to be able to call join(), you need to derive publically from Thread. 
i.e: class pseudoProcess : public Thread {
...
};