Update myshell.cc
fixed the ioRedirection with ">>". 
General problem with io redirection is that when we have ioredirection, 
we pass in the whole command line into the function we want to run. 
We are supposed to only pass in what is before the io redirection symbol.

FOR EXAMPLE: 
ls >> test.txt

We used to do:
-make stdout test.txt
- run 'ls >> test.txt' 
When the command line is run, ">>" is not a file, and the assert fails right after.

What we should do:
-make stdout test.txt
-run 'ls'